\name{grcScripts2-package}
\alias{grcScripts2-package}
\alias{grcScripts2}
\docType{package}
\title{
Genomics Resources Core Analysis
}
\description{
This package was developed for use by the University of Idaho IBEST Genomics resources core for analyzing
next-generation high-throughput sequencing data. The package utilities will take the user through data filtering, quality trimming, read overlapping, de novo assembly, and mapping reads to a reference. 
}
\details{
\tabular{ll}{
Package: \tab grcScripts2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-03\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
